requirements.txt
argparse==1.4.0
flatdict==4.0.1
interval==1.0.0
PyGithub==1.58.1
pymongo==4.3.3
python-dateutil==2.8.2
requests==2.30.0
urllib3==2.0.2
